fixed unsigned

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,9 +7,9 @@
  */
 void print_number(int n)
 {
-	unsigned long m;
-	int c;
-	unsigned long num;
+	unsigned int m;
+	unsigned int c;
+	unsigned int num;
 
 	num = n;
